[WIP] Bring in retirement buckets exercise
I created this interactive exercise in another repo before adopting
Github Pages / Jekyll / Beautiful Jekyll.

With this commit, the post containing the exercise successfully loads
the css and javascript, and is fully functional.  It's a little ugly
that the css and javascript are now loaded for all pages rather than
just the relevant post.  I think I'd have to mess with the template
structure to isolate them to just the relevant post.